Shift lib-go to using internal mockclock
Summary:
This diff makes the minimal changes required to allow lib-go to
entirely drop using benbjohnson/clock in favor of the new css.com/lib-go/clock.

The benbjohnson API returns values of *clock.Timer and *clock.Ticker that reference
it's own data structures, which ends up creating hard dependencies in projects
downstream of lib-go, so this updates those projects.

The diff doesn't really take advantage of the synchronization capabilities of the new
clock or of the ability to use time.SetSystemClock() to avoid the need to hold clock
references, leaving those more intrusive changes for separate diffs. It only aims
to disentangle from benbjohnson/clock thus opening the door for such changes in
more targeted places.

One small exception to the minimal approach is in KEQ, where I have more
familiarity so I did make some changes to take advantage of synchronization where
it was straightforward to do so, with the intent that these can help serve as examples
of how to do this in other projects. I would have done this in lib-go except no test
there actually advances the clock (outside of mockclock's own tests) so none would
serve as decent examples.

* Convert KEQ to internal mockclock
* Migrate QMS from benbjohnson/clock
* Convert cloudprint2 from benbjohnson/clock

GitOrigin-RevId: 5a51d929373712c7b1fca13c5731929ace9e454a