[keq][broadcaster] Adjust number of shards to a power of 2
Summary:
When it's not power of 2, the boundaries of e.g. 12 shards are a handful
of values off from the boundaries of 24 due to rounding differences.

This also synchronizes with broadcaster which will simplify and optimize
interactions between KEQ and broadcaster. I could sync to 12, but using
a power of 2 opens the door to later change the number of shards for both,
separately, without the rounding-error driven inefficiencies and risks since
boundaries will still align. With boundaries slightly off, one KEQ subscription
involves 3 broadcaster coordinators, one handling just a sliver of the total
range. Events on any one affect the subcription across all three, introducing
extra instability.

GitOrigin-RevId: 8b77bb561ab054843caa45c3b6d4ef3187700eee