libs-go: uuidx.Split fixes incorrect ranges when not splitting an entire UUID range
Summary:
uuidx.Split previously swallowed the error when creating a new sub range when splitting a partition. This resulted in a bunch of invalid ranges to be returned. Splitting a UUID range that didn't cover the entire UUID space was the underlying issue, where the Split function always started ranges at the equivalent `00000000-0000-0000-0000-000000000000` and didn't respect the initial range's `From` UUID.

Thanks to Kris Raney for finding the fix for this!

GitOrigin-RevId: be45dabed4ebcd614f6949a33456e07725160726