Provide a workflow library to help with fuzzing
Summary:
* [libs-go] Create generics-style heap
* [vendor-go] Add github.com/AdaLogics/go-fuzz-headers
* [lib-go] Add a library to help with fuzzing

This is implemented in support of <REDACTED> Go's native
fuzzing provides randomized data, but for many of our services (e.g. broadcaster) data is mostly
black-box and it's randomized behavior that's more interesting.

This library is targeted towards easily turning randomized data into a randomized series of actions.
The test code provides a basic example. A broadcaster fuzz test will follow that can serve as a more
robust example.

GitOrigin-RevId: d158ee35cb47380c9aed9f000b449d0d8f7fb533