[lib-go] Alter semantics of latency historgram
Summary:
First, the driving reason for this change:

* Our Java libraries provide functionality for histograms that report all latencies as a long integer representing milliseconds.
* The go metrics library normalizes all latencies to a float64 representing seconds.
* KEQ (and other projects) has both Java and Go client libraries. These now emit metrics, and we'd like the metrics to be consistent regardless of implementation language.

Since the Java implementation uses an integer, it would be quite difficult to configure it to report seconds to match Go (without losing the ability to have subsecond latencies.)

The LatencyUnit parameter on latency histograms in Go seems to allow the simple ability to report in Milliseconds. However, the setting does not have this effect. Instead,
if you set LatencyUnit to Milliseconds, it takes all of the bucket values you've given them and scales them so that they are in seconds. Then it reports
seconds.  I find this to be slightly surprising behavior.

Looking through Monorepo, it seems as though LatencyUnit: time.Millisecond is almost entirely unused. Only in
* infra/platform/log/elasticsearch/common/estools/cmd/loadtest/cmd/api.go
* infra/platform/log/elasticsearch/common/estools/pkg/loadtest/pkg.go
* infra/platform/log/kibana/es-proxy-go/pkg/proxy.go

in other words, only in one project and with most of that being in a load test, not in production code.

Given this, it seems like the best option that would allow go latencies to be reported in a matching form to java latencies is to change the semantics. The new sematics are,
* Whatever bucket values you set are used as-is. They are not scaled.
* Instead, scale the give time.Duration to the requested units - so report in seconds for time.Second (as is usual today) but report in ms for time.Millisecond

Final complication:
There is one place that uses a LatencyUnit of time.Hour. Although this isn't explictly supported (the docs say second or millisecond) it does technically work. But having
a similar semantics change for that unit wouldn't really work, the metrics system doesn't have an "hours" unit. So it should probably still be normalized to seconds, even
though this is different than what happens for Millisecond with this change in place.

GitOrigin-RevId: 62aa1466f857627976536d47d54a6159e65c0f01