[metricsx] Use the prom.DefaultRegistry instead of creating a new
Summary:
Current behaviour creates a new prometheus registry and calls the lib-go/metrics.
The metrics lib will then register the standard cpu, mem, and gc metrics on the old defaultRegistry, which is not being exportet by the metricx exporter.

This diff adds the default prom registry to the exporter created in metricsx.
Allowing for cpu, memory, and gc metrics from GO.

GitOrigin-RevId: 5d23f89125b4ae596e8978cfe26be635c614e453