[KEQ2] Support partitioned topics
Summary:
Added a new option to topic configuration - partitions. If set, this option will result in partitioning all queues into the provided number of queues. The partitioning number must be a power of two to make UUID operations efficient.

Partitioning implemented by hashing the message key using MD5 to get a new UUID, and then transforming this UUID by keeping enough bits at the beginning to be able to keep the desired number of partitions and filling the rest with zeros.

For example, `dc9076e9-2fda-4019-bd2c-900a8284b9c5` is hashed to `e5c8ce1b-7ccf-4154-a779-064e8b483edd`. With 12 bits partitioning it would be transformed to `e5c00000-0000-0000-0000-000000000000`. In this case exactly 3 characters in textual representation are retained because each character in UUID takes 4 bits. For 12 bits the topic would be partitioned into 4096 queues.

GitOrigin-RevId: b6c53b70ad6bbda813a161c68b852252a3d91caa