[KEQ2] Compute range backlog incrementally
Summary:
This change introduces incremental backlog, which tracks the subscription backlog by applying individual changes when they happen. This should remove the need to compute backlog by enumerating all queues and subscriptions periodically (with complexity of "number of queues" * "number of subscriptions").

The incremental backlog computation must happen whenever there is an update in a cursor (subscription) or a queue. To make sure the result is matching the old backlog metric, that metric is kept for some time. The internal consistency check is extended to verify the incremental backlog and compare it with the complete backlog.

GitOrigin-RevId: 9786badd336823a60a57897c6dbad31e1363b758