[WDS2] Extend ClusterMap to keep track of shards and pass them with snapshot (#49699)
When a new coordinator is created, it has an empty list of consumers. When a new consumer is connected to this coordinator, it receives a cluster map which contains no other consumers. If the consumer was previously connected to a coordinator which was re-created at a new location, it will discard the old cluster map, losing information about all other consumers (which are most likely still valid). After all consumers are connected to the new coordinator, the cluster map will be back to its state before the old coordinator was shutdown.

This change modifies the update logic of a cluster map. When a snapshot is received (first cluster message of a consumer session), the old grants assigned to a shard that have no grants in the snapshot, are kept in the cluster map. Additionally, if the old cluster map contains two grants (allocated and revoked) for the same shard and the new cluster map contains only one of these grants in a matching state, the second grant is retained.

The snapshot proto message is extended to include a list of shards. This list is used to determine a list of valid shards in the new cluster. Additionally, the assignments may have shards not listed in the list of shards; they are also included in the list of valid shards. Alternatively the update logic could be changed to enforce that assignments don't include shards not in the list of shards, but it's not clear what to do in this case: apply and log a message, or reject the update?

The primary purpose of the list of shard is to detect re-sharding, when old shards are no longer valid and all grants need to be discarded.

For backward compatibility, while service doesn't send the list of shards, this list is determined from the shards in the assignments. When this list is empty, the current grants are used to determine the list of valid shards so that old grants can be kept. This is to fix the problem described above. After this change is deployed by all the clients, this fallback will be removed.

<REDACTED>

GitOrigin-RevId: 439c9efb28bf3b7c40a0c83276be9b6b337efa25