[lib/go/tracing] Add option to support OpenTelemetry tracing and metrics initialization (#52247)
Summary:
This PR is to enable OpenTelemetry metrics and tracing instrumentation:
- Added prometheus exporter of OTel metrics. So that new code instrumented with OTel meter API will be scraped by prometheus with the same port (9090) together with existing OC metrics.
- Added environment variable `CSS_TRACING_OTEL_ENABLED`
   - Enable Span instrumented by OTel trace API to be collected by Jaeger collector.
   - Initialize Tracing OpenCensusBridge, so that all the Span instrumented by OpenCensus will be converted to OTel Spans.
- Updated the observability go Grpc sample application to verify those changes.

This PR won't impact applications if `CSS_TRACING_OTEL_ENABLED` is not explicitly set.
This PR is not ready for all the go apps to be migrated to OTEL Tracing yet due to bugs in OpenCensusBridge that trace state is not propagated: https://github.com/open-telemetry/opentelemetry-go/issues/5642.
Following up PRs will add workaround for this issue.

GitOrigin-RevId: 871abcd0fad21fcf5de7150e88d2cea8e1ebe4c1