[libs/go/backoff] Revert Fix Retry* helpers to return original values on failures (#84592)
Even though returning values with errors is something that is used occasionally (even in the standard library), we should not encourage that.

If this is required, the callers of `backoffx.Retry*` methods will have to do that explicitly.

This change brings back the original logic and adds a comment to clarify this behavior.

GitOrigin-RevId: 421db63e11916f17e72f9ff6ececf1855fe4c0a6