[KEQ2] Move scheduled messages to destination during migration (#106993)
Migration of scheduled messages is extended to move scheduled messages that are scheduled ahead of the current time. The duration is determined by `SchedulesMigrationHorizonMinutes` flagship, which default of one hour.

This moving of messages is used to detect whether there are any more scheduled messages (instead of a seek before).

Added metrics around the number of moved messages.

INF-1009

GitOrigin-RevId: 811705edfed96ff13afc1e2d8da4eb2ba30cccb1